Compute statistics updates in a queue parallel to the general update queue.
While the update queue is mostly API performance dependent, the statistics updates are mostly DB query performance dependent. Computing them in sequence resulted in a serious decline in package update check performance. With this change, statistics updates should have almost no impact on the update speed. If they turn out to be slower than the updates, they will simply be updated as fast as possible in a round-robin fashion.