Allow custom package logos & add OpenGraph
This is a split part of #265, just only for package logos

This commit goes slightly different on the caching name due to the comments, but still keeps all image files. This time the user specifies the revision of the icon and gets the default package if it is not found. If the user does not specify one they get the current logo. I decided to go for this approach due to archiving or embedding purposes, especially on google etc. where an old image of the repository might be wanted.