Add --no-serve option.
Allows to run a single poll for updates and exit, without starting up a HTTP server. This can be used in particular to let mirror updates run as an isolated process, while the web frontend runs in the background. This is necessary due to a massive memory leak triggered during the mirror update that has yet to be properly diagnosed.