fix ellipse on start page not displaying properly
Making it part of the string makes diet escape the special ampersand character. If we would insert raw HTML here we could introduce issues with HTML injection, so doing it the verbose way fixes it the easiest.

Might probably want to make this some global function instead of copy-pasting this code everywhere, but fixes the issue for now.