Avoid potential downward spiral in updatePackageList.
In case the update takes more than 30 seconds, the existing approach would potentially lead to a complete lock-up of the process, since multiple runs of updatePackageList would start to overlap. This ensures that only one update can be active at a time and that there is always a time in-between where the thread is not busy with list updates.